Team Name Regex
Teams are a little different in kore as they exist in both the database and k8s (for triggers). At the moment this leaves us open to the name being find in mysql but failing in k8s. This PR fixes the issue by ensuring the name complys with the k8s requirement (https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names). At the moment we only need to apply it here and k8s will handle the other resources for us.