Oauth Redirect URL
The currrent implementation was only setting the redirect_url in the authorize handler. This meant when the request wen across a pod, you'd send an invalid token exchange as the redirect_url wouldn't have been set. I've kept the functionality of using the host headers as this is quiet useful if your reverse proxying things.