Fixing the missing jwks_uri issue
This came to light when using Okta as an IDP.
When given an OpenID URL the openid-client will setup an Issuer.
If the URL contains `.well-known/openid-configuration` then it will load the config from that endpoint.
If, as in our case the URL set in the config did not contain `.well-known/openid-configuration`, it will load the config from one of `/.well-known/openid-configuration` or `/.well-known/oauth-authorization-server`, depending on which one responds first (no joke)
The latter endpoint, `/.well-known/oauth-authorization-server`, did not have the same config ie. it was missing the `jwks_uri`
This meant that it was pretty much down to luck as to whether the Issuer was setup with the `jwks_uri` or not.
The Issuer is configured once per restart which is why restarting appeared to fix the issue

As a fix for this I've added some additional logic around the `KORE_IDP_SERVER_URL` env var that's used in the config, to ensure it either already contains `.well-known/openid-configuration` or it is appended.