Cluster visibility and editability through the UI.
Addresses #494 - Being able to drill-down into the detailed status of the cluster and it's sub-resources.
Addresses #493 - Being able to delete a cluster in a failed state.
Addresses #567 - Being able to edit an existing cluster's configuration after it is built through the UI.

Introduces the concept of 'fixed' plan parameters defined in the plan schema. These are parameters which
cannot be edited after a cluster is created, such as region.