VPC Deletion
Currently we suffer from a number of issues when trying to delete the VPC. Though we are not the only ones

- https://github.com/weaveworks/eksctl/issues/1325
- https://github.com/aws/amazon-vpc-cni-k8s/issues/608
- https://github.com/terraform-providers/terraform-provider-aws/issues/11473

This issues appears to be related to the AWS CNI which isn't releasing the ENI. At present the only way to fix the bug is but manually deleting the ENI and the security group, removing the dep and allowing the VPC to delete. This PR trys to codify that fix, removing any flacky ENI and security groups. Note it doesn't remove everything attached to the VPC, so if they've attached additonal security groups not tagged as defined those will be left as blockers.

The terraform issue (https://github.com/terraform-providers/terraform-provider-aws/issues/11473) appears to suggests an ordering issue with the worker nodes being stripped of the policy before they have the chance to clean up; however we are not presently deleting these so i'm not sure if that's just something related to that module which experienece a similar issue