Allow to query generated service plan details with one API call
There is a new api endpoint: GET /serviceplans/{name}/details?team=team_name&cluster=cluster_name

If the plan has empty schemas, then it will return the schemas from the service kind.

With the optional team and cluster parameters you can generate the final schemas and the plan configuration for a given cluster context.

When we create a new service, this API endpoint should be called with the right context, so we can fill in the right cluster related parameters.

Example:
An S3 plan has a configuration parameter `target_account_id` with a value of `{{ .cluster.status.providerData.awsAccountID }}`.
If the team/cluster is passed to the API call, the value reference will be replaced with the real AWS account ID (on the schemas and on the plan configuration too).