Namespace Claim Imports
This is a first pass at imports - long term we probably want something more elaborate. Perhaps
a locked (i.e. k8s lock) service which connects out and watches the namespace events? though I'm
not keen on the idea of a long term connection like that .. Or an agent model where by we send
back events to kore which a service and pick off topic and handle?